refactor(code-review): move socialShareButtons to constants
